snapshot-agent: move the GPU-CR artifact sweeper into its own gpucr package
The sweeper, group-store helpers, and shared GPU-CR env/const surface
move from the unbounded utils package into pkg/snapshot-agent/gpucr,
which names and bounds what the package may contain. pod-utils stays
put. StartGPUCRSweeper drops its package-stuttering prefix and becomes
gpucr.StartSweeper.
