+++ pkg/snapshot-agent/gpucr/gpucr_internal_test.go
@@ -1,4 +1,4 @@
-package utils
+package gpucr
 
 import (
 	"fmt"
